[fix] pdate BellIcon file path and rename component
